+++ src/REvoDesign/basic/__init__.py
@@ -12,7 +12,7 @@ from .data_structure import IterableLoop
 from .extensions import FileExtension, FileExtensionCollection
 from .group_registries import GroupRegistryItem
 from .menu_item import MenuCollection, MenuItem
-from .param_toggle import ParamChangeRegister, ParamChangeRegistryItem
+from .param_toggle import ParamChangeRegistryItem
 from .plugin_registry import PluginRegistry, build_plugin_registry
 
 __all__ = [
@@ -23,7 +23,6 @@ __all__ = [
     "FileExtensionCollection",
     "GroupRegistryItem",
     "ParamChangeRegistryItem",
-    "ParamChangeRegister",
     "MenuItem",
     "MenuCollection",
     "PluginRegistry",